connection: cleanup @connection_set_uuid_error test and fix leak
First of all, the test leaks "con_con2" connection. It means, if
you run it twice in a row, it fails. Fix that.

Also, the 'with options "|| true"' relies on the fact that "Execute
reproducer" will use command_code(), which uses a shell and that
the options are just concatenated as strings "bash $SCRIPT $OPTIONS"
That's too many assumptions. Don't do that. As always, the reproducer
script should exit with zero on success and fail otherwise. No need for
this hack.