tags:nmstate_upstream_setup: avoid finding wrong connection to clone
This whole setup seems ugly and fragile.

With loopback branch there will be generally a "connected (externally)"
profile on "lo", so

  context.process.nmcli("-g NAME con show -a").strip()

will give "testeth0\nlo", which cases a failure:

  Exception: `nmcli con clone testeth0
  lo nmstate` returned exit code 10
  STDOUT:

  STDERR:
  Error: unknown connection 'testeth0
  lo'.

Ignore externally-connected profile. that at least filters out "lo".
It's still ugly. The test setup has no concept of what state to expect or
ensures that the right state is present. Instead, it just looks ad
`nmcli con show -a` and hopes that the returned profile is the right
one.