nmci.util: always create new _Timeout instance in nmci.util.start_timeout()
_Timeout.loop_sleep() has an internal flag so that "while
timeout.loop_sleep():" will always iterate at least once.

If we pass the timeout instance down (to a called function), and somebody
calls start_timeout() on it, we want that the loop-sleep state of the
caller's timeout is not affected and the caller of start_timeout() gets
an instance with reset loop_sleep() behavior.

The solution is to not reuse the _Timeout instance, just use it to
get the remaining time.