nmci.util: drop "default_timeout" argument from nmci.util.start_timeout()
It's not useful.

For one, None already had a meaning of infinity. Now it was hijacked to
mean "use-default" value.

But when the caller has no "real" value, they don't need to call:

  nmci.util.start_timeout(default_timeout=d)

they can just call:

  nmci.util.start_timeout(timeout=d)

On the other hand, if you got a timeout instance from somewhere, then
that's the timeout that should be honored.

Every caller that passes a timeout can specify whatever it wants. And,
every implementation can themselves have default values that they apply.
There is no need for this special timeout value that indicates
"default".