nmci.util: add "at_least_once" argument to timeout.loop_sleep()
The idea with "loop_sleep()" is two fold:

1) on the first iteration, we would not sleep and run the loop
  right again.
2) even if the timer already expired at the point when starting
  the iteation, it would still iterate at least once.

The idea of 2) is that you may want to run the loop at least once. For
example, if we want to assert for something (with polling and retry),
we want to run the check at least once (even if the timer already
expired).

But maybe 2) is not always best. Add an option to opt out of that.