contrib: rename "contrib/gi/{libnm_snapshot_=>}checkpoint.py"
The file is imported directly from NetworkManager. There is has
the name "checkpoint.py". It is confusing, to keep the file with a
different name, because then it's harder to see where it came from.

Also, "contrib/gi/checkpoint.py" is really a small script that has
several commands to interact with checkpoints. That this uses libnm
internally, does not seem so important that it needs to be part of the
name. Also, "snapshot" is a redundant word to "checkpoint". Meaning,
it's not clear that the longer name is better.

Rename.