MPTCP support: prepare step and associated cleanup
MPTCP in NM depends on respective sysctl settings. Therefore
infrastructure is added that allows to save and restore state of:
  * sysctls matching pattern (previous commit)
  * current configuration of MPTCP exposed by iproute2

Adding the step 'Prepare simulated MPTCP setup ...' that creates simple
environment suitable for MPTCP testing with other endpoints in a custom
"mptcp" namespace. Number of interfaces, their prefix and a mode of
endpoints in "mptcp" namespace are configurable.