version_control: support interval ranges
This is longly desired feature, to support "holes" in version intervals.
So, now every version case should be possible to describe in
single scenario, no need to copy paste scenario and adjust the tags.

Example usage:
```
@ver+=1.20
@ver-1.31.8
@ver+=1.42
```

This means, feature added in 1.20 was broken in 1.31.8
(and test would fail there), and was fixed again in 1.42.

All unit tests pass as they used to, so this patch should not break
existing `@ver` tags.