nmci: better handle races in nmci.nmutil.connection_show()
nmci.nmutil.connection_show() calls `nmcli connection show` and parses the
output. nmcli ensures that this output is self-consistent and represents the
correct state at a certain point in time. Afterwards, the function augments
the result with additional information, by fetching more data from D-Bus.
That data is fetched at a different time, so there is no guarantee that the
result is consistent.

Avoid this problem by fetching twice. If the result differs, we suspect
a race and retry. Obviously, this has an overhead, as we now take twice
as long as before.