nmci.nmutil: don't mutate and construct new list in nmci.nmutil.device_status()
It's ugly to do:

  result = [_device_update_ipaddrs(d) for d in result]

Either _device_update_ipaddrs() is a pure function that does not
modify the input and creates a new output. Then it makes sense to
generate a new list, by mapping each element to something new.

Or _device_update_ipaddrs() modifies the input, but then there is no
need to generate a new list. There is also no need to return anything
from the function.

Do the latter.