runtest: adjust running_NM_version_check() to check version of NetworkManager in path
With "nm-in-container.sh" script, we install NetworkManager from the source tree
in "/sbin/NetworkManager" and restart the daemon. Basically, we overwrite the
files from the RPM, and the version reported by `rpm -q NetworkManager` is
not the version that is expected to run.

We expect that the running version is the same as the installed
`NetworkManager -V`, not what rpm says.