prepare.py: MPTCP: workaround for failing Fedora builds
NM currently doesn't remove MPTCP endpoints for an interface that goes
down (eth0 for MPTCP cases) which makes existing checks for MPTCP
endpoints wrong on systems that have MPTCP enabled by default (Fedora).

A workaround is to disable MPTCP, restart NM and flush the endpoints
early in the Prepare so that the rest of scenario works exactly the same
as on systems with MPTCP disabled.