service: don't use reset=False for nmci.nmutil.restart_NM_service()
It doesn't seem to make sense. When NM-ci restarts the daemon,
it doesn't want that systemd fails and says we restarted too
fast. We always want to reset the fail counter.

Otherwise, if you run a test with "Restart NM" in a loop, and
the test is fast, it starts failing after a few iterations.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1304