nmci: autogenerate docs from docstrings
Use sphinx-markdown-builder module, which generated .md file. It
creates `nmci/doc/index.md` file, which is linked to `nmci/README.md`
and that is rendered by GitLab once in `nmci` directory.

Is is not possible to include in Gitlab markdown, so `README.md` is
rewritten to RestrucutredText - `nmci/doc_src/index.rst`, which is
source for the documentation.