make testeth0 recreation more robust: vethsetup
We rely on /tmp/testeth0 being present for outgoing connection
recreation, yet we don't take precautions neither to ensure that the
file is there, nor to back it up as a part of potentially destructive
scenarios.

This commit makes vethsetup's check() create new /tmp/testeth0 when it
is missing or empty.

MR: https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1321