cleanup.py: add CleanupFileRestore class
The current CleanupFile class only deletes a given list of files. Let's
add a complementary class that will back the content of a specified file
up and restore it at the clean up stage.

The RestoreFile() class as of now:
  * neither does error handling in _add_cleanup() nor in _do_cleanup()
    stage
  * saves all files' contents as distinct objects even no matter if the
    file at a given path was already backed up
Fixing either of these will require addition of skips to _add_cleanup().

Also, the contents of the file is stored in memory so this feature is
not suitable for (potentially) large files

MR: https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1321