nmci.util: add nmci.util.random_generator() for fuzzying
Add a new mechanism for stable/reproducible random numbers.

We should have a way to randomize tests. For example, tests should
randomly configure `[main].plugins=keyfile` or
`[main].plugins=ifcfg-rh`. Most of our tests should work with both
configurations, but it does not scale to run the entire CI twice, with
both settings plugins. Well, basically we already do that, as we run CI
on rhel8 and rhel9, but rhel8 is going to get phased out and it doesn't
scale. In particular, we may have another setting where we randomly
either use "[main].dhcp=internal" or "[main].dhcp=dhclient". The point
is, there are countless of interesting combinations to test, but we
cannot cover them all explicitly (either, because we don't think about
explicitly testing that combination, and also we don't have the
resources/time to cover them all).

Add nmci.util.random_generator(seed), which you can call to get a
random generator. The "seed" should be fixed and can be arbitrarily
chosen, but for every use case that you want to fuzz, choose a unique(!)
seed. The seed is then combined with a global seed.

The per-request seed (which is hard-coded in the test) is combined with
a global seed.  The global seed can be set via the NMCI_RANDOM_SEED
environment variable. That is in particular useful for testing or
reproducing a particular fuzz configuration. We will also log
NMCI_RANDOM_SEED, so you can find the seed after a run and set it
manually. If NMCI_RANDOM_SEED is unset, we honor ".tmp/rnd_global_seed".
If that file doesn't exist, we generate a random seed and persist it in
".tmp/rnd_global_seed".

https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1359