features: support setting the DHCP range in prepare_simdev() to one IPv4 address
prepare_simdev() accepts an ipv4 parameter, for the DHCP range. However,
previously that always had to be a prefix like "192.168.99". Then the
method would configure {ipv4}.1 as IP address on the interface and set
the DHCP range to {ipv4}.10 to {ipv4}.15.

That means, the actually leased IP address is not deterministically
known by the test.

Extend that, to accept a complete IPv4 address. In that case, the range
just contains this single IP address. And the interface gets the ".1"
address.