connection: don't check for read-only output in @connection_describe
The read-only property is special. Unlike most properties of a profile,
it's not user configuration. Instead, it's something that the daemon
exposes about the profile.

Consequently, it could not be set in nmcli.

Also, the read-only flag has no more meaning for the daemon. That's
because when a profile is loaded from a read-only location (like
"/usr/lib/"), then the profile can still be modified and NetworkManager
will store an override in "/etc".

It was dropped from nmcli. Which is an API change, but something that
hopefully nobody cares about.

Just don't test for this. Also don't test on old versions, because the
describe output is entirely generated. By testing just for a few
properties, we can verify that the general mechanism works. It would be
unlikely to break the describe output for "read-only" but not other
properties.

See-also: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1567