general: add tests for handling autoconnect/ignore-carrier with bonds
Add 3 tests related to autoconnect/ignore-carrier handling.

Note that the current behavior of NetworkManager is still faulty. Hence,
the tests only check the current behavior (as far as we get). What we
want, is later to fix NetworkManager and also notice and adjust the
tests. Basically, these tests are all xfail.

https://bugzilla.redhat.com/show_bug.cgi?id=2156684
https://bugzilla.redhat.com/show_bug.cgi?id=2180363

https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1419