nmci: add option to synchronously reload in nmutil.reload_NM_service()
Sending a SIGHUP signal is asynchronous (meaning, after the signal was
sent, we don't know whether the change was already processed by
NetworkManager). For that there is a D-Bus call, which is supposed to
only return after the changes where done. We could either make that call
via busctl, or call `systemctl reload NetworkManager`. The effect is
basically the same, because the systemd reload just does the same busctl
call (see NetworkManager.service file).