nmci: don't wait for NetworkManager in nmutil.reload_NM_service()
reload() only makes sense, if NM is already running (the command should
fail if that were not the case).

Note that if NetworkManager is not running, pkill likely will already
fail. So it's unclear why the "wait_for_nm_bus()" is useful.

Nonetheless, still call self.wait_for_nm_bus() as we did.  Just don't
give any additional poll interval. Check once, and if NetworkManager
is not ready, fail. The polling is unnecessary or would hide a bug.