nmci/tests: fix test_process_run() if the NM-ci directory is a symlink
Python's os.getcwd() and /bin/pwd resolve symlinks (/usr/pwd defaults to "-P").
Unlike shell's pwd, which would not do that.

So if you enter the NetworkManager-ci directory via a symlink, the the test fails.

This test is probably rather fragile due to that, but with this fix it should work
well enough.