nmci: add nmci.nmutil.reboot_NM_service()
This contains the code from the "Reboot" step.

The code was only accessibly from inside "features/steps/service.py",
but not from other features (unless we would introduce a way to import
the "service.py" step file).

The steps themselves should be very simple, and the reusable code should
be in nmci.

Move the code.