nmci: fix timeout handling for nmci.nmutil.wait_for_nm_bus()
wait_for_nm_bus() polls, so the timeout argument controls how long we poll.
But each individual check (the busctl call) is not subject to the same
timeout.

Consider if you run nmci.nmutil.wait_for_nm_bus(0), then that should
mean to only check once without retry. It does not mean, that the busctl
must complete in no time (which it cannot).