general: fix race in @clean_device_state_files test
Previously, the test recreated an interface 120 times in the background,
while checking that the number of device state files oscillates. That
seems a rather racy approach.

Interestingly, I don't see this test fail in the stats. However, I have
a merge request for NetworkManager, where this fails frequently (I don't
really understand why it fails there so often).

Just do it differently. What NetworkManager is doing is that after
writing 100 times to the device state files, it starts a prune action
and will delete files that refer to interfaces that no longer exist.

So we can just create 120 times the same device (creating thereby a lot
of ifindexes that come and go). Afterwards check that the number of
files is below that number, and assume that this is because files were
cleaned.

Also, create those interfaces faster and only sleep 0.5 seconds between
interfaces. The test takes a lot of time otherwise.