general: fix @nmcli_shows_correct_routes test to only check specific device
NM-ci should run on various systems. For example, when running
NetworkManager in a container with "nm-in-container" script, then there
is an unmanaged "tap0" device that connects to the host. Those routes
are not shown in `nmcli` output.

But beside the special "nm-in-container" case, I had this test failing,
when a previous test failed to cleanup the interfaces, and there were
left-over interfaces with IPv6 link local addresses. Sure, that failure
helped to find that other issue, but it seems not best to write a
fragile test, even if that can uncover some other issues.