vlan: refactor cleanup of add_bridges_vlans_range() step
Don't use tags for cleanup. Use cleanup actions which are scheduled
right when we do the thing that requires the cleanup.

Otherwise, the step 'Add bridges over VLANs in range from...' and the
@remove_vlan_range are separate pieces, when they really belong
together.

Also, rework how the cleanup is done. First, stop the NetworkManager
service, then delete all files and interfaces, then restart
NetworkManager.

Use CleanupCallback for that, and register a list of callbacks. The
benefit of that is that each callback can fail individually, and the
remaining cleanup actions are still run. This could have also be done by
scheduling multiple cleanup actions, and tweak the priority accordingly.
This approach ensures that all the step of the cleanup are bundled
together.