vlan: fix cleanup for @vlan_create_1000_bridges_over_1000_vlans test
A change in NetworkManager will make it autoconnect more eagerly.
Deleting the interfaces first (via iproute2), can lead to a new
autoconnect of the vlan interfaces. Ensure we really get rid of those
interfaces, by doing it twice.

This is a hack to get the test passing. It will be refactored later.