nmci.cleanup: drop "args" from nmci.cleanup.cleanup_add()
Python has many useful aspects of functional programming. In particular,
it's trivial to define functions and closures.

Don't support a mechanism, where callback can have an arbitrary number
of **kwargs. Instead, callback must always be a callable without
arguments. If you want to pass arguments, create a closure.

I think this

     nmci.cleanup.cleanup_add(
         callback=nmci.ip.link_set,
         name=f"link-up: {orig_name}",
         unique_tag=(orig_name, True),
         args={"ifname": orig_name, "up": True},
     )

is worse than

     nmci.cleanup.cleanup_add(
         callback=lambda: nmci.ip.link_set(ifname=orig_name, up=True),
         name=f"link-up: {orig_name}",
         unique_tag=(orig_name, True),
     )

Don't add too many ways, how the callback can be specified. Especially,
later, we will support a list of callbacks. That will be useful,
to ensure that each element of the list is called even in face of
exceptions. Such a list of callbacks would be odd to combine with a
separate args argument. Just use lambda, as intended.