nmci.cleanup: split CleanupCallback from Cleanup base class
The Cleanup base class should only have the necessary parts to track the
cleanup action and handle them. It should not also provide an
alternative to subclassing, where a callback can be provided.

It almost never makes sense to both subclass and provide a callback.
Instead, add a CleanupCallback subclass, which can be used for exactly
the cases where you want to provide a callback.

Next, the callback will be more complex, so adding that complexity to
the base class would be wrong. It's already wrong, that every Cleanup
base instance would have a callback argument which none of the existing
subclasses care about (because they override _do_cleanup_impl()).