nmci.cleanup: rename Cleanup._do_cleanup() method
There will be a hook called Cleanup._also_needs_impl(), which is what
subclasses should override. That one needs the "impl" suffix, because
we already have "self._also_needs" which is something else. To be consistent,
rename Cleanup._do_cleanup() to Cleanup._do_cleanup_impl(), to follow
the pattern and have a unique name.