nmci.cleanup: support list of callbacks in CleanupCallback
The cleanup will aim to call all callbacks in face of an error, and
re-emit any exception at the end.

This makes it more convenient to register a list of callback actions,
while being robust. It is always a test failure, if a callback raises an
exception. However, we want to ensure that a test failure does not
prevent followup callbacks to be executed as they are important for
cleanup.