bond: adjust tags for @bond_set_balance_slb_options test
The test is currently failing ([1]), which requires nftables fixes.
However, a workaround was merged to NetworkManager ([2]) in 1.43.6+
and 1.42.7+, so the test should start passing without [1].

Adjust the version tag to run only on those versions.

Also, drop the @skip_in_centos tag. It's not clear why it wouldn't work
in CentOS, given the other dependencies are suitable.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2177667
[2] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/d3b54963622f242db1ebeda21dedd9558b484355