steps: rewrite skip step
Having the general skipper not bound to RPM is better. Also, skipping in
scenarios is probably more prefered, since the conditions of the
environment are directly visible in the report, and not hidden in tags
or prepare shell scripts.

Step name is confusing, maybe we should come up with some better name,
but it would be good if `skip` word was still contained in the sentence.

Also `* Skip if next step passes:` might work too, but it is harder to
implement, since we would have to tell behave that the failure of the
next step should be treated as pass (not easy to do, probably changing
some atributes of runner object would be necessary not to stop execution
of the next steps).