run: fix installation of CentOS 9 Stream due to nmstate build dependencies
NetworkManager requires certain build dependencies, like libndp-devel
and teamd-devel.  Those source packages are not available in plain
RHEL9/C9S, so we have a copr repository that rebuilds those packages
([1]).

The problem is, in copr there is only an epel9/centos9 release, but
there are no specific builds for rhel-9.2, rhel-9.3, etc. Hence, we
currently have (for example) libndp-1.7-6 build for epel9.

If you then start centos:stream9:

  $ podman run --privileged -ti centos:stream9
  # dnf install 'dnf-command(copr)' -y
  # dnf copr -y enable nmstate/nm-build-deps
  # dnf install libndp-devel -y

then we get libndp-1.7-6 from copr.

A subsequent `dnf upgrade` fails then with

  # dnf upgrade -y
  Last metadata expiration check: 0:00:31 ago on Mon May 22 14:45:10 2023.
  Error:
   Problem: package libndp-devel-1.7-6.el9.x86_64 requires libndp = 1.7-6.el9, but none of the providers can be installed
    - cannot install both libndp-1.8-4.el9.x86_64 and libndp-1.7-6.el9.x86_64
    - cannot install the best update candidate for package libndp-devel-1.7-6.el9.x86_64
    - cannot install the best update candidate for package libndp-1.7-6.el9.x86_64

Avoid that with the "--nobest" option. We just want a suitable libndp.

[1] https://copr.fedorainfracloud.org/coprs/nmstate/nm-build-deps/