rename NetworkManager.conf snippets
NetworkManager.conf snippets in conf.d directories are loaded by sorting
the file by names. It's common practice to name them something like
"20-connectivity-fedora.conf".

NM-ci should not hijack the very highest priority. First of all, plain
NetworkManager ships with few configuration snippets (00-server.conf and
20-connectivity-fedora.conf). Those don't tend to conflict with what
NM-ci does, because they inherently have a low priority (low numbers).
Also, NM-ci can and should be aware of the configuration snippets that
are present in the test (if they are installed by packages on the tested
system -- contrary to being created by a human developer who debugs a
test). It is a known fact, that there are no snippets present with a
relevant high priority, which would be required to be overwritten with a
priority "99-".

"99-*.conf" is the very highest priority, unless the user starts
hacking the naming scheme with "a0-even-higher-prio.conf". So 99-*.conf
does not ensure somebody could not hack the scheme, it just makes
it ugly to do so.

Also, using "99-*.conf" for various files, leads to oddities like
"99-xtest*.conf" and "99-xxcustom.conf". We should not take away
the highest numbers, if we then still plan to use even higher numbers.

Rename those files to "95-*.conf". It's higher than anything which we
expect there. If a user wants for manual testing override that, they
still have some space (99-*.conf).

Also add a "nmci" part to the file name, so we can easily see where this
file comes from.