openvpn: enable openvpn_persist test for rhel9
The cleanup step for removing openvpn connection is unnecesary and in
RHEL9 it might be buggy as we did a connection reload right before it.

This is not a NetworkManager issue, the assert happens when openvpn
takes too much to recover and regenerate tun1 interface. Probably
NetworkManager could be improved but it seems a corner case that users
won't hit.

https://bugzilla.redhat.com/show_bug.cgi?id=2217022