general: fix @nmcli_device_wifi_with_two_devices
This test was executing a suite of python unit test, and then
parsing all the results output to get only one of the results. The
test has started to fail because a (expected) warn message is printed
now, breaking the parsing logic. Also, the output from unittest is not
exactly the same in all versions of python, like python3 in Fedora37,
that prints some more details, breaking also the parsing logic.

Execute only the test we are interested in, not the others, and use
the returned value instead of parsing the output.