vrf: fix failures with IPv4 DAD enabled
It's fine to have the same address on different interfaces in separate
VRFs. It's not fine if those interfaces are connected to veths
attached to the same bridge ("inbr" in "vethsetup" namespace), because
the ARP request on eth1 is looped back to eth4 and vice versa.

Instead of eth1 use eth10, which is connected to a different bridge
"simbr".