cloud: fix race in cloud_setup_keep_secondary_ips
Since both IPv4 and IPv6 are enabled in conX1 and both are
may-fail=yes, it might happen that IPv6 completes before
IPv4. Therefore, when "nmcli connection up" returns, the DHCPv4
address might not be configured yet. In that case, when we add an IPv4
address externally, the resulting order of addresses is not as
expected and the test fails. This is visible especially when IPv4 DAD
is enabled by default, because it delays DHCPv4.

Make IPv4 may-fail=no to fix this race condition.