ipv4:dhcp: add new tests for client-id "none" and default
A new special value "none" is being introduced for ipv4.dhcp-client-id
in NM 1.45.5. It works with dhclient and internal client. Add new tests
to verify its behaviour.

Also, add 2 new tests for the default value of ipv4.dhcp-client-id when
using dhclient. The default behaviour is to use the client-id if it's
defined in dhclient.conf, or don't send any client-id otherwise.