ipv4:dhcp: skip ipv4_dhcp_client_id_default_dhclient_unset in RHEL8
In RHEL8, dhclient turns the option `-i` on by default, so the default
behaviour when no dhcp-client-identifier is specified is different: it
will send the DUID as client-id, instead of sending none as usual.

Just add @rhelver+=9 since we also have @ver+=1.45.5 and we are not
going to test that version in RHEL 7, anyway.