dracut: skip remote connection var2 test on el8
This test started to fail after fixing renew and rebind timeouts, only on rhel8.
Scenario where each interface is set separately works fine, problem seems to be
with multiconnection.
Since there is workaround for this, and rhel9 is not affected (NM as service in dracut),
it is safe to skip.