dispatcher: add tests for generic connections with device-handler
Test the new feature that allows to define a 'device-handler' for a
generic connection. This device handler is a dispatcher script that
gets called by NM to add and delete the interface for the generic
connection. In this way, it becomes possible to support new kinds of
devices by writing an appropriate script, without patching NM.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1752

https://issues.redhat.com/browse/RHEL-1567