steps, nmci: add steps and function to update keyfiles
The newly created nmci.misc.keyfile_update() function and accompanied
steps update the given configfile/keyfile with the configuration given
as a string with the same format.

Note: the function and steps do not handle backups or cleanups in any
way, they should only be used for ephemeral files or for permanent
changes to the system.