ipv4+ipv6: skip the tests checking the empty address for ipv[46].addresses
Because we are allowing configuring the static routes when the addresses
are empty since NM version 1.45.10, the error message for validating the
`ipv[46].addresses` also changes slightly. Therefore, in some tests
where checking the error message for validating the `ipv[46].addresses`,
we want to intentionally skip those tests after NM version 1.45.10.

https://issues.redhat.com/browse/RHEL-5098
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1832