ipv4: add test for ipv4.dhcp-dscp
Test that the values passed via `ipv4.dhcp-dscp` take effect. Note
that the property only works with the internal client at the
moment.

Test the 3 supported values: CS0 (default), CS4, CS6 (previously the
default).

Check that the right DSCP value is set in: the DHCP DISCOVER, the DHCP
REQUEST and when renewing the lease, as this last case uses a
different socket (UDP instead of packet socket).

https://issues.redhat.com/browse/RHEL-16040