ipv4: make @dhcp_internal_nak_in_renewing scenario more robust
As of now, the scenario can fail in a racy way, if periodic check for
172.25.1.200 address presence on testX4 interface falls into the short
period between IP removal after DHCP NAK and its reconfiguration a
fraction of second later.

With this commit, we wait until NM gets the NAK plus some time to get
the address again and only then we check the address' presence.

https://issues.redhat.com/browse/RHEL-26777