tags: restore resolv.conf symlink to resolved after @dns_dnsmasq
When systemd-resolved is active before the scenario, tag @dns_dnsmasq
removes /etc/resolv.conf and stops the service; after the scenario it
just restarts the service.

Therefore, if resolv.conf was a link to
run/systemd/resolve/stub-resolv.conf before the scenario, the link is
not restored after the scenario and NetworkManager will write to
resolv.conf as a plain file.

Instead, we need to restore the previous state of the file.

We already assume in tag @restore_resolvconf that when
systemd-resolved is active, resolv.conf should be a symlink to the
resolved stub file. Do the same in @dns_dnsmasq.