libreswan: fix DNS test
The NM behavior is slightly different depending on whether it is
writing directly to resolv.conf or it is using a backend that supports
split DNS.

As documented in the "ipv4.dns-priority" section of "man nm-settings",
when writing directly to resolv.conf NM just lists all the name
servers with increasing priority. When using split DNS, NM applies
some post-processing to the DNS and domain lists; the effect is for
example that the name server from a VPN with the default route
"shadows" the name server from the Ethernet connection, which is
ignored.

To avoid this ambiguity always run the test with systemd-resolved.

https://issues.redhat.com/browse/RHEL-29302