tags: mask systemd-resolved in @dns_default
Currently the tag only stops systemd-resolved; since the unit is D-Bus
activated it can be started again when its D-Bus API is called for
example by NetworkManager when resolving an address.

I think the @dns_default tag should test a scenario without
systemd-resolved, therefore mask the service so that it's not
restarted automatically.