simwifi: make p2p tests more reliable
Adjust fixed timeouts where possible.
The initial `sleep 2` after `wpa_supplicant` start is important.
The rest of the waits is dynamic.
Also, do not start `wpa_supplicant` child in verbose mode with `-d`,
it never responds for some reason, even if the number of lines it
outputs is low (maybe 200).