publish_behave_logs: get rid of the delay
The delay used to be needed to avoid copying empty report files. These
days, reports are successfully copied even when the delay is gone.

A great benefit is that when your run one scenario in a loop with no
sleep:
    for {0..$n}; do run/runtest.sh $scen; done
You'll no longer lose all the reports from the loop but the last one,
all of them will be accessible via http://host:8080/