contrib/cloud-setup-mock: put/get actual resources. Add OCI provider.
Instead of being a simple map of path->string, let's make it a bit more
sophisticated so it stores actual Python objects and convert them
from/to JSON.

This is an advantage because changes to "children" resources don't
require to edit the parent paths. For example, if a new VNIC is added to
the path ".../macs/<MAC>/", there is no need to also edit the path
".../macs" to update the list of VNICs that it returns.

This is specially useful if we use JSON responses instead of plain text
ones. In this case, a GET to ".../macs" returns all the information of
all the VNICs at once. Because of that, if we change a single field of a
single VNIC, we would need to update all the parent paths to reflect
that field's change (i.e. it would affect to the content of all the
paths .../macs, .../macs/<MAC>, .../macs/<MAC>/ipv4,
.../macs/<MAC>/ipv4/ipAddresses, .../macs/<MAC>/ipv4/ipAddresses/0,
etc.)

Also, add OCI (Oracle Cloud) new provider which we want to support in
nmcs.