contrib/cloud-setup-mock: add OCI default resources with VLAN
The most complex scenario that we need to test nm-c-s for OCI is a
baremetal instance with some VNICs requiring VLAN configuration.

Baremetal is detected if nicIndex field is present. VLAN configuration
is required when vlanTag != 0. Provide default resources with this
configuration. NMCI test cases use --empty so they are not affected, but
this is useful for development and debugging.