ipv4,bridge: don't use IP configs on ports
Setting ipv4 or ipv6 configs on port devices is not valid. Nmcli raises
an error if you try, and the daemon silently drops that IP config during
the verification phase.

Because of a bug in nmcli, it was possible to send a port connection
with an IP config if the type and/or port-type configs were passed last,
because arguments where processed in strict order. In this case, nmcli
didn't raise any error but the daemon dropped it.

The nmcli bug is being fixed by
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/2165
so now a few tests has started to fail. Remove the IP configs. They were
useless, anyway.