nmcli: add tests for nmcli arguments order and port-type guessing
nmcli was not processing correctly the type and port-type if they were
specified at the end of the arguments list. Now it's fixed by
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/2165.

For OVS it was necessary to specify the port-type, even if it is
implicit by the connection type. This is fixed by the same MR.

Add tests for both things.