bridge: test that autoconnect-ports works with virtual device ports
Sometimes, under specific circumnstances, autoconnect-ports may fail if
the port is a virtual device that hasn't been realized yet. For example,
a VLAN that doesn't actually exist because it's connection is not
activated yet, so the device hasn't been created.

The specific circumnstances are:
- Port with autoconnect no
- Controller with autoconnect-ports yes

This started to happen after NM's commit 774badb1519a ('core: prevent the
activation of unavailable devices'). It is being fixed now by
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/2203
and this test guards against regressions on this.