dracut: disable checksum on net devices
New kernel (6.15+) does not compute checksum on virtual devices (bridge). However, virtio device is not considered virtual inside VM, so it expects checksum to be computed. This inconsistency is ignored when getting the lease (expected), but not during renew (checksum should match). Hence, disable checksum on virtio device in qemu.