sriov: add test for the "sriov.preserve-on-down" property
The new "sriov.preserve-on-down" property controls whether
NetworkManager preserves the SR-IOV parameters set on the device when
the connection is deactivated, or whether it resets them to their
default value. The SR-IOV parameters are those specified in the
"sriov" setting, like the number of VFs to create, the eswitch
configuration, etc.

Add a test.