netdevsim: use upstream patch to set permanent address
Apply new patch from upstream kernel, if it is not applied already.
We have to use new_port and new_device sys files to add new devices.
For now, MAC is hardcoded in the script, can be changed if needed.
This supports up to 255 devices (last part of mac), which is far more than enough.

https://issues.redhat.com/browse/NMT-1847