gsm_sim: replace modem name from modemu to smd0
In rawhide, the generic plugin is failing for some reason. After searching in MM src, generic plugin is complaining it can not load drivers. Not sure what it means, but I have noticed  `smd0` device is treated as virtual and it does skip this driver check. I see no  side effect or regression, and it is working on rawhide too now.